Configure to use the PREFIX environment variable
The Docker installation can now use the PREFIX environment variable to
set the location that the application will expect to be served from.

For example, setting PREFIX="" (the default), the application will
expect to be served from the URL http://hostname/. If instead you set
the PREFIX="/observations", the application will expect to be served
from the URL http://hostname/observations/.

This is especially useful when using the Docker container to serve just
one section of a larger website.